LayoutTests/fast/forms/color/input-color-disabled-expected.txt

Tests that it is not possible to select a color on a input type=color that is marked as disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS onChangeCount is 0
PASS onChangeCount is 0
PASS input.value is "#000000"
PASS successfullyParsed is true

TEST COMPLETE
